Added the UserActivateExtend class
